requirements.txt
appdirs==1.4.0
Django==1.8.17
djangorestframework==3.5.3
olefile==0.44
packaging==16.8
Pillow==4.0.0
Pygments==2.2.0
pyparsing==2.1.10
six==1.10.0
